Add support for custom tag resolution in react render (#496)
* Add support for custom tag resolution

Currently, the way in which the react render resolves
tag names is hardcoded to a conventional, but not required within react
default, of react components always being named with an uppercase
letter.

This change allows for users to change how the resolution of tag names
to react components occurs. It does so be creating new options to both the
dynamic and static react renderers that allows the user to pass a function that overrides the default `tagName` function.

As an example, without this change, if I want to style certain "default" nodes, like the
table components (thead, tr, etc), then I must override all of the nodes
and then set a name with uppercase letter (like `THead`) and then pass a matching component. This requires a lot more ceremony to use markdoc if the user wants to rely on components instead of stylesheets to change default components.

The motivation for this change comes from how MDX allows the user to
also [pass components](https://mdxjs.com/docs/using-mdx/#components)
but allows for any arbitrary component, so in some MDX implementations,
it becomes a common practice to just define a map of *all* components.
See [the following
example.](https://github.com/shadcn-ui/ui/blob/main/apps/www/components/mdx-components.tsx#L41)

 As an alternative, the existing logic *could*
change to *first* check the component repo for a matching name instead
of requiring it to be an uppercase letter, but changing that now could
be a subtly breaking behavior. This commit instead opts for introducing
a new option to the renderers, making this change opt-in.

* fix styles