Update formatter regex for list items and headings (#582)
Fixes a small bug with the markdoc formatter where certain strings were incorrectly escaped as markdown characters. For example `#not a heading` would get escaped as `\#heading` or `*Not a list item` would get escaped to `\*Not a list item` when they should be maintained. Both `>blockquote` and `> blockquote` are valid blockquotes.

[Commonmark playground to demonstrate examples](https://spec.commonmark.org/dingus/?text=%23%20is%20a%20heading%0A%0A%23%23is%20not%20a%20heading%0A%0A-%20is%20a%20list%20item%0A%0A-is%20not%20a%20list%20item%0A%0A%3Eis%20a%20blockquote%0A%0A%3Eis%20also%20a%20blockquote)